+++ src/data/orderbook.rs
@@ -43,10 +43,10 @@ pub fn encode_orderbook(orderbook: &OrderBookMsg) -> OrderbookResult<Vec<u8>> {
     // This data should have "at least" 21 bytes.
     let mut bytes = Vec::<u8>::with_capacity(21);
 
-    // 1. 交易所时间戳: 8 字节
+    // 1. 交易所时间戳: 6 字节
     bytes.extend_from_slice(&ExchangeTimestampRepr(orderbook.timestamp).to_bytes());
 
-    // 2. 收到时间戳: 8 字节
+    // 2. 收到时间戳: 6 字节
     bytes.extend_from_slice(&ReceivedTimestampRepr::try_new_from_now()?.to_bytes());
 
     // 3. EXCHANGE: 1 字节
@@ -99,10 +99,10 @@ pub fn encode_orderbook(orderbook: &OrderBookMsg) -> OrderbookResult<Vec<u8>> {
 pub fn decode_orderbook(payload: &[u8]) -> OrderbookResult<OrderBookMsg> {
     let mut reader = BufReader::new(payload);
 
-    // 1. 交易所时间戳: 8 字节时间戳
+    // 1. 交易所时间戳: 6 字节时间戳
     let exchange_timestamp = ExchangeTimestampRepr::try_from_reader(&mut reader)?.0;
 
-    // 2. 收到时间戳: 8 字节时间戳 (NOT USED)
+    // 2. 收到时间戳: 6 字节时间戳 (NOT USED)
     reader.consume(8);
     // let received_timestamp = ReceivedTimestampRepr::try_from_reader(&mut reader)?;